// frontend/src/models/enums.ts
export enum NodeType {
    Pump = 'pump',
    Sump = 'sump',
    Tank = 'tank',
    Bore = 'bore',
    Govt = 'govt',
}

export enum ProductType {
    EvaraTank = 'evaratank',
    EvaraFlow = 'evaraflow',
    EvaraDeep = 'evaradeep',
    None = 'none'
}

export enum NodeStatus {
    Normal = 'Normal',
    Warning = 'Warning',
    Critical = 'Critical',
}